app/requirements.txt
pandas
Flask
mysql-connector-python
sqlalchemy
pymysql